+++ src/main/java/io/kontur/eventapi/dao/mapper/FeedMapper.java
@@ -67,7 +67,8 @@ public interface FeedMapper {
                       @Param("eventDetails") String eventDetails,
                       @Param("enriched") Boolean enriched,
                       @Param("episodes") String episodes,
-                      @Param("name") String name);
+                      @Param("name") String name,
+                      @Param("enrichmentAttempts") Long enrichmentAttempts);
 
     Integer getNotEnrichedEventsCount();
 }