+++ src/main/resources/db/changelog/v0/feed_data_enrichment_index.sql
@@ -1,5 +1,6 @@
 --liquibase formatted sql
 
---changeset event-api-migrations:feed_data_enrichment_index runOnChange:false
+--changeset event-api-migrations:feed_data_enrichment_index runOnChange:true
 
-create index feed_data_updated_at_feed_id_is_not_enriched_idx on feed_data (updated_at, feed_id) where not enriched;
+drop index if exists feed_data_updated_at_feed_id_is_not_enriched_idx;
+create index feed_data_updated_at_feed_id_is_not_enriched_idx on feed_data (enrichment_attempts, updated_at, feed_id) where not enriched;